rename some files for consistency. split p96_rtg.c into p96gfx_card.c for the card function wrappers and p96gfx_rtg.c for the rtg support code. inline some card access operations.
git-svn-id: https://svn.aros.org/svn/aros/trunk@56056 fb15a70f-31f2-0310-bbcc-cdcc74a49acc
