+++ AROS/arch/m68k-amiga/hidd/p96gfx/p96gfx_rtg.h
@@ -374,21 +374,46 @@ struct ModeInfo {
 #define P96BoardType_UAEGfx                         14
 #define P96BoardType_Vampire                        25
 
+/* inline access wrappers */
+static inline APTR gp(UBYTE *p)
+{
+    return ((APTR*)p)[0];
+}
+static inline ULONG gl(UBYTE *p)
+{
+    return ((ULONG*)p)[0];
+}
+static inline UWORD gw(UBYTE *p)
+{
+    return ((UWORD*)p)[0];
+}
+static inline void pp(UBYTE *p, APTR a)
+{
+    ((APTR*)p)[0] = a;
+}
+static inline void pl(UBYTE *p, ULONG l)
+{
+    ((ULONG*)p)[0] = l;
+}
+static inline void pw(UBYTE *p, WORD w)
+{
+    ((WORD*)p)[0] = w;
+}
+static inline void pb(UBYTE *p, BYTE b)
+{
+    ((BYTE*)p)[0] = b;
+}
+
+/* RTG Support functions */
+extern const UBYTE modetable[16];
+
 WORD getrtgdepth (ULONG rgbformat);
 ULONG getrtgformat(struct p96gfx_staticdata *csd, OOP_Object *);
 void makerenderinfo(struct p96gfx_staticdata *csd, struct RenderInfo*, struct bm_data*);
 struct ModeInfo *getrtgmodeinfo(struct p96gfx_staticdata *csd, OOP_Object *sync, OOP_Object *pixfmt, struct ModeInfo *modeinfo);
+void InitRTG(APTR boardinfo);
 
-APTR  gp(UBYTE *p);
-ULONG gl(UBYTE *p);
-UWORD gw(UBYTE *p);
-void pp(UBYTE *p, APTR a);
-void pl(UBYTE *p, ULONG l);
-void pw(UBYTE *p, WORD w);
-void pb(UBYTE *p, BYTE b);
-
-extern const UBYTE modetable[16];
-
+/* Card functions */
 BOOL FindCard(struct p96gfx_staticdata *csd);
 BOOL InitCard(struct p96gfx_staticdata *csd);
 BOOL SetDisplay(struct p96gfx_staticdata *csd, BOOL state);
@@ -404,7 +429,7 @@ GetFeatureAttrs
 SetFeatureAttrs
 */
 
-/* Render Operation Stubs .. */
+/* Card Render Operation Stubs .. */
 BOOL DrawLine(struct p96gfx_staticdata *csd, struct RenderInfo *ri,
     struct Line *line, ULONG rgbformat);
 BOOL BlitRect(struct p96gfx_staticdata *csd, struct RenderInfo *ri,
@@ -430,14 +455,13 @@ BOOL SetSpritePosition(struct p96gfx_staticdata *sd);
 BOOL SetSpriteImage(struct p96gfx_staticdata *sd);
 BOOL SetSpriteColor(struct p96gfx_staticdata *sd, UBYTE idx, UBYTE r, UBYTE g, UBYTE b);
 
-/* real RTG only functions */
+/* Real RTG only card functions */
 ULONG GetPixelClock(struct p96gfx_staticdata *csd, struct ModeInfo *mi, ULONG index, ULONG rgbformat);
 ULONG ResolvePixelClock(struct p96gfx_staticdata *csd, struct ModeInfo *mi, ULONG pixelclock, ULONG rgbformat);
 ULONG SetClock(struct p96gfx_staticdata *csd);
 void SetMemoryMode(struct p96gfx_staticdata *csd, ULONG rgbformat);
 void WaitBlitter(struct p96gfx_staticdata *csd);
 void SetInterrupt(struct p96gfx_staticdata *csd, ULONG state);
-void InitRTG(APTR boardinfo);
 /*
 TODO:
 GetVSyncState
