+++ AROS/arch/m68k-amiga/hidd/p96gfx/p96gfx_startup.c
